requirements.txt
torch>=2.2.0
transformers>=4.57.0
accelerate>=1.0.0
opencv-python-headless>=4.10.0
Pillow>=10.0.0
PyYAML>=6.0.0
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0
safetensors>=0.4.0
sentencepiece>=0.2.0
qwen-vl-utils>=0.0.10
torchvision>=0.17.0
